fix: remove old Article component file
