+++ src/routes.js
@@ -8,8 +8,7 @@ import routesConst from './constants/routes'
 import statusCodeConst from './constants/status-code'
 // components
 import FallbackPage from './containers/ServiceWorkerFallbackPage'
-// import Article from './containers/Article'
-import Article from './containers/Article-new.js'
+import Article from './containers/Article.js'
 import SystemError from './components/SystemError'
 // lodash
 import get from 'lodash/get'
